segmentation/01-sam-transformer-fine-tuning/requirements.txt
transformers
datasets
monai
opencv-python
pillow
matplotlib